requirements.txt
torch
transformers
datasets
tqdm
matplotlib
seaborn
numpy
pillow
path
python-dotenv
langchain
langchain-openai
bs4
langchain-chroma
langchain-community
langchain-core
langchain-text-splitters
lxml



